channels change JS into TS
