+++ src/pages/about/channels/index.tsx
@@ -1,29 +1,38 @@
 import React, { useState, useEffect } from 'react';
 import { connect } from 'dva';
 import moment from 'moment';
-import { history } from 'umi';
+import { Dispatch, history } from 'umi';
 import { Table, Button, Badge, Space, Modal } from 'antd';
 import { ExclamationCircleOutlined } from '@ant-design/icons';
-import { Breadcrumb } from 'components';
+import { Breadcrumb } from '@/components';
 import CreateChannelModal from './components/CreateChannelModal';
-import { MenuList, getCurBreadcrumb } from 'utils/menu';
-import baseConfig from 'utils/config';
-import { Roles } from 'utils/roles';
+import { MenuList, getCurBreadcrumb } from '@/utils/menu';
+import baseConfig from '@/utils/config';
+import { Roles } from '@/utils/roles';
 import { ChannelStatus } from './_config';
+import { ConnectState } from '@/models/connect';
+import { TableColumnsAttr, } from '@/utils/types';
+import { ChannelSchema} from '@/models/channel';
 
 const breadCrumbItem = getCurBreadcrumb(MenuList, '/about/channels');
 
-function ChannelManagement(props) {
+export interface ChannelManagementProps {
+  dispatch: Dispatch;
+  location: Location;
+  qryLoading: boolean;
+  User: ConnectState['User'];
+  Channel: ConnectState['Channel'];
+}
+function ChannelManagement(props: ChannelManagementProps) {
   const { dispatch, location, qryLoading = false, User, Channel } = props;
   const { networkName, userRole } = User;
   const { channelList, channelTotal } = Channel;
-  const [columns, setColumns] = useState([]);
   const [pageNum, setPageNum] = useState(1);
   const [pageSize] = useState(baseConfig.pageSize);
   const [createChannelVisible, setCreateChannelVisible] = useState(false);
 
   // 获取通道列表
-  const getChannelList = (current) => {
+  const getChannelList = (current: number) => {
     const offset = ((current || pageNum) - 1) * pageSize;
     const params = {
       offset,
@@ -39,7 +48,7 @@ function ChannelManagement(props) {
   };
 
   // 翻页
-  const onPageChange = (pageInfo) => {
+  const onPageChange = (pageInfo: any) => {
     setPageNum(pageInfo.current);
     getChannelList(pageInfo.current);
   };
@@ -50,15 +59,16 @@ function ChannelManagement(props) {
   };
 
   // 关闭 创建通道弹窗
-  const onCloseCreateChannel = (res) => {
+  const onCloseCreateChannel = (res: any) => {
     setCreateChannelVisible(false);
     if (res) {
+      //TODO:没有传参数
       getChannelList();
     }
   };
 
   // 点击 查看组织
-  const onViewOrg = (record) => {
+  const onViewOrg = (record: ChannelSchema) => {
     history.push({
       pathname: `/about/channels/organizationList`,
       state: { ...record },
@@ -66,7 +76,7 @@ function ChannelManagement(props) {
   };
 
   // 点击 查看节点
-  const onViewPeer = (record) => {
+  const onViewPeer = (record: ChannelSchema) => {
     history.push({
       pathname: `/about/channels/nodeList`,
       state: { ...record },
@@ -74,7 +84,7 @@ function ChannelManagement(props) {
   };
 
   // 点击 查看合约
-  const onViewContract = (record) => {
+  const onViewContract = (record: ChannelSchema) => {
     history.push({
       pathname: `/about/channels/chaincodeList`,
       state: { ...record },
@@ -82,7 +92,7 @@ function ChannelManagement(props) {
   };
 
   // 点击 查看详情
-  const onViewDetail = (record) => {
+  const onViewDetail = (record: ChannelSchema) => {
     history.push({
       pathname: `/about/channels/channelDetail`,
       state: { ...record },
@@ -90,9 +100,9 @@ function ChannelManagement(props) {
   };
 
   // 点击操作按钮, 进行二次确认
-  const onClickToConfirm = (record, type) => {
+  const onClickToConfirm = (record: ChannelSchema, type: any) => {
     let tipTitle = '';
-    let callback = null;
+    let callback = () => { };
     switch (type) {
       case 'enable':
         tipTitle = '启用';
@@ -108,6 +118,7 @@ function ChannelManagement(props) {
     Modal.confirm({
       title: 'Confirm',
       icon: <ExclamationCircleOutlined />,
+      //TODO:model上的ChannelSchema上没有channelName这个属性
       content: `确认要${tipTitle}通道 【${record.channelName}】 吗?`,
       okText: '确认',
       cancelText: '取消',
@@ -115,70 +126,66 @@ function ChannelManagement(props) {
     });
   };
 
-  // 用户身份改变时，表格展示改变
-  useEffect(() => {
-    const data = [
-      {
-        title: '通道名称',
-        dataIndex: 'channelId',
-        key: 'channelId',
-      },
-      {
-        title: '通道别名',
-        dataIndex: 'channelAliasName',
-        key: 'channelAliasName',
-      },
-      {
-        title: '通道状态',
-        dataIndex: 'channelStatus',
-        key: 'channelStatus',
-        render: (text) =>
-          text ? (
-            <Badge
-              color={ChannelStatus[text].color}
-              text={ChannelStatus[text].text}
-              style={{ color: ChannelStatus[text].color }}
-            />
-          ) : (
-            ''
-          ),
-      },
-      {
-        title: '通道描述',
-        dataIndex: 'channelDesc',
-        key: 'channelDesc',
-        ellipsis: true,
-      },
-      {
-        title: '创建者',
-        dataIndex: 'createUser',
-        key: 'createUser',
-      },
-      {
-        title: '创建时间',
-        dataIndex: 'createdAt',
-        key: 'createdAt',
-        render: (text) => moment(text).format('YYYY-MM-DD HH:mm:ss'),
-      },
-      {
-        title: '操作',
-        key: 'action',
-        width: '22%',
-        render: (text, record) => (
-          <Space size="small">
-            <a onClick={() => onViewOrg(record)}>查看组织</a>
-            <a onClick={() => onViewPeer(record)}>查看节点</a>
-            <a onClick={() => onViewContract(record)}>查看合约</a>
-            <a onClick={() => onViewDetail(record)}>详情</a>
-          </Space>
+  const columns: TableColumnsAttr[] = [
+    {
+      title: '通道名称',
+      dataIndex: 'channelId',
+      key: 'channelId',
+    },
+    {
+      title: '通道别名',
+      dataIndex: 'channelAliasName',
+      key: 'channelAliasName',
+    },
+    {
+      title: '通道状态',
+      dataIndex: 'channelStatus',
+      key: 'channelStatus',
+      render: (text) =>
+        text ? (
+          <Badge
+            color={ChannelStatus[text].color}
+            text={ChannelStatus[text].text}
+            style={{ color: ChannelStatus[text].color }}
+          />
+        ) : (
+          ''
         ),
-      },
-    ];
-    setColumns(data);
-  }, [userRole]);
-
+    },
+    {
+      title: '通道描述',
+      dataIndex: 'channelDesc',
+      key: 'channelDesc',
+      ellipsis: true,
+    },
+    {
+      title: '创建者',
+      dataIndex: 'createUser',
+      key: 'createUser',
+    },
+    {
+      title: '创建时间',
+      dataIndex: 'createdAt',
+      key: 'createdAt',
+      render: (text) => moment(text).format('YYYY-MM-DD HH:mm:ss'),
+    },
+    {
+      title: '操作',
+      key: 'action',
+      width: '22%',
+      render: (text, record: ChannelSchema) => (
+        <Space size="small">
+          <a onClick={() => onViewOrg(record)}>查看组织</a>
+          <a onClick={() => onViewPeer(record)}>查看节点</a>
+          <a onClick={() => onViewContract(record)}>查看合约</a>
+          <a onClick={() => onViewDetail(record)}>详情</a>
+        </Space>
+      ),
+    },
+  ]
   // 页码改变、搜索值改变时，重新查询列表
   useEffect(() => {
+    //TODO:没有传参
     getChannelList();
   }, [pageNum]);
 
@@ -219,7 +226,7 @@ function ChannelManagement(props) {
   );
 }
 
-export default connect(({ User, Layout, Channel, loading }) => ({
+export default connect(({ User, Layout, Channel, loading }: ConnectState) => ({
   User,
   Layout,
   Channel,
